moved load_file to utils
