+++ distributed/utils.py
@@ -11,12 +11,16 @@ import os
 import re
 import shutil
 import socket
+from importlib import import_module
+
 import six
 import sys
 import tblib.pickling_support
 import tempfile
 import threading
 import warnings
+
+from .compatibility import cache_from_source, invalidate_caches, reload
 
 try:
     import resource
@@ -696,3 +700,38 @@ def open_port(host=''):
     port = s.getsockname()[1]
     s.close()
     return port
+
+
+def load_file(path):
+    """ Loads modules for a file (.py, .pyc, .zip, .egg) """
+    directory, filename = os.path.split(path)
+    name, ext = os.path.splitext(filename)
+    names_to_import = []
+    if ext in ('.py', '.pyc'):
+        if directory not in sys.path:
+            sys.path.insert(0, directory)
+        names_to_import.append(name)
+        # Ensures that no pyc file will be reused
+        cache_file = cache_from_source(path)
+        if os.path.exists(cache_file):
+            os.remove(cache_file)
+    if ext in ('.egg', '.zip'):
+        if path not in sys.path:
+            sys.path.insert(0, path)
+        if ext == '.egg':
+            import pkg_resources
+            pkgs = pkg_resources.find_distributions(path)
+            for pkg in pkgs:
+                names_to_import.append(pkg.project_name)
+        elif ext == '.zip':
+            names_to_import.append(name)
+
+    loaded = []
+    if not names_to_import:
+        logger.warning("Found nothing to import from %s", filename)
+    else:
+        invalidate_caches()
+        for name in names_to_import:
+            logger.info("Reload module %s from %s file", name, ext)
+            loaded.append(reload(import_module(name)))
+    return loaded
